fix circular includes
